util/types, column, expressions: use types.Convert to replace Col.CastValue
The two methods has similar functionality and has complex logic, after this change,
the code base would be easier to maintain.

For `CAST` function, the behaviour is different than the general purpose `types.Convert`,
so I implemented a dedicated function for it.

+++ expression/expressions/unary.go
@@ -142,9 +142,9 @@ func (u *UnaryOperation) Eval(ctx context.Context, args map[interface{}]interfac
 		if err != nil {
 			return types.UndOp(a, op)
 		} else if n == 0 {
-			return int8(1), nil
+			return int64(1), nil
 		}
-		return int8(0), nil
+		return int64(0), nil
 	case opcode.BitNeg:
 		a := Eval(u.V, ctx, args)
 		if a == nil {
